pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-gateway"
version = "0.1.0"
description = "MCP gateway/proxy with logging, denylist, and caching"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
  "aiohttp>=3.9",
  "pyyaml>=6.0",
  "psycopg[binary,pool]>=3.1",
  "orjson>=3.10",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[project.scripts]
gateway = "mcp_gateway.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
